Data 클래스 dynamic 키워드로 런타임 오퍼레이터 연산 & int, float 암시적 형변환

+++ Assets/Scripts/Entity/Guardian/Guardians/Weapon.cs
@@ -33,7 +33,7 @@ public abstract class Weapon : MonoBehaviour
 
     public WeaponType WeaponType;
 
-    public float AttackDamage;
+    public int AttackDamage;
 
     public bool isEquiped => guardianData != null;
